kernel-rebase: add indication of conflicting fixups
Add indication of fixups which need manual resolution.
Currently when a fixup conflicts then there is only
a messsage displayed "Conflict found" and triage/rebase
continues. The indication will be in the form of a+b
in the "need manual resolution" column, for example
"4+2" means that 4 commits and 2 fixups need manual
resolution.

BUG=None
TEST=Run continuous rebase triage and verify that number
     of fixups which require manual resolution is correct

Change-Id: I5df899a89567cff25bdd73c0636d53b3348ce941
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/3845356
Tested-by: Łukasz Bartosik <ukaszb@google.com>
Commit-Queue: Łukasz Bartosik <ukaszb@google.com>
Reviewed-by: Guenter Roeck <groeck@chromium.org>