fflash: Use in-memory ext4 instead of tmpfs
Since R107-15071.0.0 on linux-5.10, or R107-15072.0.0 on linux-5.15,
our way of making a tmpfs with exec no longer works.
I suspect this is due to hardening efforts in CL:3852537 and CL:3834992.

The kernel seems happy with in-memory ext4.
Use that instead as a workaround.

BUG=None
TEST=fflash $cherry_running_R107_15071
TEST=fflash $strongbad_running_R107_15072

Change-Id: I1f0f2fa33ee6dda4d7a13434cdf66e6b11e6df42
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/3862125
Tested-by: Li-Yu Yu <aaronyu@google.com>
Reviewed-by: Terry Cheong <htcheong@chromium.org>
Commit-Queue: Li-Yu Yu <aaronyu@google.com>