cros_unibuild_convert: Allow for a 2-phase probe DUT -> generate migration
When working with a couple of DUTs, it's convenient to first generate
the DUT data, save it to a file, and then actually run the migration
later with the saved data.  Especially when leasing from Skylab where
you may only have an hour to use the DUT online.

This could also be useful for possibly writing tests for this script
... which might not be necessary given I intend to throw away this
script once we migrate the last 4 boards to unibuild.

BUG=b:187793386
TEST=./cros_unibuild_convert.py --dut-data-out=/tmp/duts.json \
           -d ${DUT_IP} -b lars
     ./cros_unibuild_convert.py -d /tmp/duts.json -b lars --dry-run

Change-Id: I0624e34cdbc05a2f7339bf9ac21cc7c1438b6b15
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/3885103
Reviewed-by: Eizan Miyamoto <eizan@chromium.org>
Tested-by: Jack Rosenthal <jrosenth@chromium.org>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>