labtunnel: Add support for making btpeer tunnels with wificell cmd
Wificells can have btpeers, but only the btpeer cmd can tunnel to
btpeers. This changes abstracts out the btpeer tunneling logic into
a common function and moves the chameleond port flag up a level so
that both the btpeers and wificell commands can make tunnels to
btpeers in the same way. The wificell cmd has a new --btpeers flag
to specify the amount of btpeers to tunnel to, which defaults to 0
to maintain current default behavior.

This prevents the need to run multiple labtunnel commands when
trying to create tunnels to both the wifi devices and btpeers in a
wificell at the same time, which is a common use-case for
coexistence tests.

BUG=b:250005351
TEST=build passes; btpeers cmd tested; wificell cmd tested with
and without new --btpeers flag and is working as expected

Change-Id: I948d1d651a321193d18ae857a6b935507dc2006f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/4029816
Tested-by: Jared Bennett <jaredbennett@google.com>
Reviewed-by: Shijin Abraham <shijinabraham@google.com>
Commit-Queue: Jared Bennett <jaredbennett@google.com>