labtunnel: Refactor version reporting to use native Cobra versioning
The Cobra CLI framework labtunnel uses has support for specifying
the version of the CLI when declaring the command. Labtunnel currently
has a separate version command to display the version. This change
removes the version command in favor of using the Cobra versioning
support.

BUG=None
TEST=successfully builds; calling "labtunnel -v" displays the version
as expected

Change-Id: I53120fd1e6fc2f0ed9c1249eeeece1460aed689b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/4026321
Tested-by: Jared Bennett <jaredbennett@google.com>
Commit-Queue: Jared Bennett <jaredbennett@google.com>
Reviewed-by: Shijin Abraham <shijinabraham@google.com>