labtunnel: Add support for --tauto flag for tauto tunnel differences
Tunnels for btpeers and chameleon devices are made directly to the
chameleond service running on them at port 9992. This is how Tast uses
these devices, but actually Autotest/Tauto connects to them via ssh
and from there they will tunnel separately.

This adds a new flag, -a or --tauto, that a user can provide to
indicate that the tunnels they are creating are for Tauto and should be
created in the way that Tauto expects. That means for btpeer and
chameleon device tunnels, the remote port is for ssh nort chameleond.
If any other tunneling differences are found, this same flag will be
the only thing a user has to specify, rather than specifying flags for
each individual difference.

BUG=None
TEST=builds successfully; validated chameleon and btpeer tunnels use
the remote ssh port when the new flag is set

Change-Id: If996c0ab5e4962a82ae8ee2d494d827e23788736
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/4029818
Commit-Queue: Jared Bennett <jaredbennett@google.com>
Reviewed-by: Shijin Abraham <shijinabraham@google.com>
Tested-by: Jared Bennett <jaredbennett@google.com>