openwrt: Add iputils-ping utility to openwrt image
wifi.APSupportedRates and wifi.PowerSave failing on openwrt due to non-integer interval parameter passed into ping command. OpenWRT's busybox ping package doesn't support this. wifi.WMM is also impacted by this package as the -Q flag is not supported on openwrt's ping package. We add iputils-ping package to provide a package which supports these functions.

We also modify the environment so that when the ping command is invoked, the correct package is used

BUG=b:258551363
TEST=wifi.APSupportedRates, wifiPowerSave, and wifi.WMM pass on openwrt on openwrt images with ping while failing to pass before the change. This change only affects openwrt devices.

Change-Id: I12db73965dbe85f9ae8b50eb3def9544339c5e0f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/4032762
Commit-Queue: Billy Zhao <billyzhao@chromium.org>
Tested-by: Billy Zhao <billyzhao@chromium.org>
Reviewed-by: Jared Bennett <jaredbennett@google.com>