labtunnel: Add new hosts subcommand
The logic to create a tunnel for a specific type of device has been
refactored into common functions called by each subcommand. This
made possible the new "hosts" subcommand, added in this CL, that
allows you to create many different types of tunnels to arbitrary
hosts, including hosts that are just IPs, as defined by a series of
CLI flags.

Since this also slightly changes the output of labtunnel (each tunnel
type has an ID in the name now), the examples in the README have been
updated with output from this new version. Additional examples have
been added for the new "hosts" subcommand as well.

BUG=b:250005351
TEST=builds successfully; tested each command by re-running the
examples in the README, all working as expected

Change-Id: I92306b48786ce768a1a518207c01c5878c4bbba9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/4035304
Commit-Queue: Jared Bennett <jaredbennett@google.com>
Reviewed-by: Shijin Abraham <shijinabraham@google.com>
Tested-by: Jared Bennett <jaredbennett@google.com>