kernel-rebase: add initial support for alternative bisection strategy
This commit adds -a/--alternative switch to bisect_branch.py which will
provide the different way of bisecting and verifying the kernel.

Currently it doesn't do much, but it's a ground work for future
improvements. The alternative flow is split in two steps:

  1. Rebase commits from repo tag stored in variable "end" onto commits
     up to tag in variable "begin" but pick Chromium patches only, then
     try to rebuild the kernel.
  2. Get list of upstream commits between "begin" and "and", split the
     list in chunks (currently only two halves) and then apply them on
     top of previous changes. For every chunk of patches, verify that
     it's corrently building and fail fast if it does not.

In future, the alternative strategy will grow a bit, as well as support
flashing the kernel builds on DUTs during bisection to verify they're
booting properly.

BUG=b:260831489
TEST=./bisect_branch.py

Change-Id: I9311c345e87ffe09d64d6cff5b0ec2cafba54eba
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/4075576
Commit-Queue: Łukasz Bartosik <ukaszb@google.com>
Tested-by: Łukasz Bartosik <ukaszb@google.com>
Reviewed-by: Łukasz Bartosik <ukaszb@google.com>
Reviewed-by: Guenter Roeck <groeck@chromium.org>