Add CTPBuilder object and validation
This object will be used to schedule builds eventually, this just
establishes the struct and validation. The final usage should look like:
c = CTPBuilder{...}
c.ScheduleBuild()

Because CTPBuilder is a large struct with many optional fields, I opted
to export the struct directly, and plan to call the validation in the
ScheduleBuild methods. This is opposed to some method called
NewCTPBuilder() which handles the validation/defaults as a constructor

BUG=259101455
Test Plan=`go test ./...`

Change-Id: I6928e5f8ac9cd5af3ed8f909ac9dc7b211052341
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/4081121
Reviewed-by: Jard Loucks <jaredloucks@google.com>
Commit-Queue: Eli Trexler <elijahtrexler@google.com>
Tested-by: Eli Trexler <elijahtrexler@google.com>