Add Buildbucket client
Some users of this tool might not want to use this client; specifically
crosfleet has the logic to generate a client with a specific user agent.

We can allow users to call ScheduleBuild with or without a buildbucket
client. If they provide a client, we use that. Otherwise, we create this
client for users. That allows users who don't care about this user agent
(like PVS) to not worry about making a buildbucket client, while
providing appropriate backwards compatibility to crosfleet

BUG=b/259101455
Test Plan=`go test ./...`

Change-Id: I59c0f6d6d802069def86699cda3a6c6062e1c6e9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/4085525
Reviewed-by: Jard Loucks <jaredloucks@google.com>
Commit-Queue: Eli Trexler <elijahtrexler@google.com>
Tested-by: Eli Trexler <elijahtrexler@google.com>