fflash: Add bazel build support
Compared to the hand crafted ninja, this shortens the build when
editing a non dut-agent file from 6s to 1s, mostly due to ninja
being timestamp based, and we xz the dut-agent even when it is
unchanged. However a clean build is slower due to needing to
compile protoc.

The ninja build file is unchanged for now.

1.  Update go deps.
2.  Use rules_go and gazelle to populate BUILD files.
3.  Add custom build rules for embedded agent.

BUG=None
TEST=bazel test //...
TEST=bazel build //...
TEST=bazel run //cmd/integration-test -- $DUT

Change-Id: Iaa8c0f9a48e958b356b294f18973c7f33372ebe3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/4194894
Reviewed-by: Terry Cheong <htcheong@chromium.org>
Commit-Queue: Li-Yu Yu <aaronyu@google.com>
Tested-by: Li-Yu Yu <aaronyu@google.com>