fflash: Prefer branched versions on the same milestone (R###)
Sometimes a milestone is not branched on the latest major version
on that milestone, for example R109-15236.x.0 while there is
R109-15237.0.0.

Update GetLatestBuildWithPrefix() to pick:
1.  Latest milestone
2.  Branched versions
3.  Latest major version
4.  Latest minor version
among multiple matches.

Note that maxVersion() was implemented incorrectly. It passed something
like "R109-15237.0.0" to semver.Compare. The usage of semver is removed
as we have to parse the version string ourselves anyway.

BUG=b:266721499
TEST=bin/fflash $DUT -R109  # flashes R109-15236.80.0 instead of R109-15237
TEST=bin/integration-test $DUT
TEST=go test ./...

Change-Id: I3759453d517ed0f7522449794970ca13bfd9927b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/4194895
Commit-Queue: Li-Yu Yu <aaronyu@google.com>
Reviewed-by: Brett Brotherton <bbrotherton@google.com>
Tested-by: Li-Yu Yu <aaronyu@google.com>