fflash: Use LATEST file to find version to flash for -R###
Sometimes the latest build directory for a given milestone
is not complete: the directory exists, but the artifacts are not
available yet. Use the LATEST- file for milestones to hopefully
find a complete version.

Add a --dry-run flag to make testing this easier.

Also fix up how we refer to version numbers in fflash.
The new names are mostly inspired from /etc/lsb-release:
> CHROMEOS_RELEASE_BRANCH_NUMBER=80
> CHROMEOS_RELEASE_BUILD_NUMBER=15236
> CHROMEOS_RELEASE_CHROME_MILESTONE=109
> CHROMEOS_RELEASE_PATCH_NUMBER=0
> CHROMEOS_RELEASE_VERSION=15236.80.0

BUG=b:266721499
TEST=bin/fflash dut1 --dry-run -R111
TEST=bin/fflash dut1 --dry-run -R110
TEST=bin/fflash dut1 --dry-run -R109
TEST=bin/fflash dut1 --dry-run -R108
TEST=bin/fflash dut1 --dry-run
TEST=bin/integration-test dut1

Change-Id: I1498c958f9fa58f4d6d7f9b11ef4aff785859e5f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/4194900
Commit-Queue: Li-Yu Yu <aaronyu@google.com>
Reviewed-by: Brett Brotherton <bbrotherton@google.com>
Tested-by: Li-Yu Yu <aaronyu@google.com>