fflash: Ignore incomplete versions
Sometimes a directory for a version is created on gs:// but the
artifacts are not yet available. Check for them.

This is required to get the canary branch i.e. main
to work with -R properly. Other release branch should already be
covered by CL:4194900.

BUG=b:266721499
TEST=bin/fflash dut1 --dry-run -R111

Change-Id: Iee982318c4bb29199b3abcb9ee011b0f25d7c15f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/4194901
Tested-by: Li-Yu Yu <aaronyu@google.com>
Reviewed-by: Brett Brotherton <bbrotherton@google.com>
Commit-Queue: Li-Yu Yu <aaronyu@google.com>