testplan: Output CoverageRules instead of HWTestPlans.
- Currently when testplan is run in non-CTP1 mode,
VMTestPlans are not allowed. This is confusing to
users, and in the future we may want to have CQ base
collect behavior on which builds have {HW|VM}TestPlans,
so this restriction will need to be removed.

- Outputting two different protos ({HW|VM}TestPlans to
the same file would be confusing and potentially
cause issues when callers parse the file. These protos
are basically just CoverageRules anyway, so output
CoverageRules instead.

BUG=b:267805660
TEST=CQ
TEST=Manually ran testplan.

Change-Id: I003da81efbada3e997b57abe166aa3cc4e35b0d9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/4209164
Auto-Submit: Andrew Lamb <andrewlamb@chromium.org>
Reviewed-by: Navil Perez <navil@google.com>
Commit-Queue: Andrew Lamb <andrewlamb@chromium.org>
Tested-by: Andrew Lamb <andrewlamb@chromium.org>