Chameleon: Generate a script that generated igtrc onto the DUT
[Why]
IGT required ~/.igtrc to find and run chamelium tests.
It's a hassle to ask every developer to understand how to do that.

[How]
Automate this process by only requiring the IPs and the script will
generate the file and put it in the right place on DUT.

BUG=b:267194166
TEST=./generate_igtrc.sh -d fwd -c 192.168.0.29

Change-Id: I203d14df598f50c593206e86e2660edfed0b8303
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/4219336
Reviewed-by: Ilja Friedel <ihf@chromium.org>
Tested-by: Mark Yacoub <markyacoub@google.com>
Reviewed-by: Drew Davenport <ddavenport@chromium.org>
Commit-Queue: Mark Yacoub <markyacoub@google.com>