local-cft: Light weight test finding
Local-cft's test-finding currently uses the workflow that starts cros-dut and requires a host to be provided. However, test finder does not interact with cros-dut there is no need to launch it or provide a host.

Implement the -findtest flag to have lighter weight use

BUG=b:267787002
TEST=./local-cft -findtest ...

Change-Id: Ia3812b050d623dd3ba8d5bfedd4a68cec0674249
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/4219701
Tested-by: Chris DeLaGarza <cdelagarza@google.com>
Commit-Queue: Chris DeLaGarza <cdelagarza@google.com>
Tested-by: Derek Beckett <dbeckett@chromium.org>
Reviewed-by: Derek Beckett <dbeckett@chromium.org>