kernel-rebase: add move disposition overwrite
Add 'move' disposition overwrite. When we had a need to move
a commit to another topic branch we used kind of a hack. We
dropped the commit on its source topic branch and then pick
it up as a fixup on destination topic branch. The' move'
overwrite disposition simplifies this operation and makes
it cleaner.

BUG=b:268175572
TEST=run kernel continuous rebase triage

Change-Id: I7251bfc72e23e17cb20f0309a40c76c4cfb65081
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/4225221
Tested-by: Łukasz Bartosik <ukaszb@google.com>
Reviewed-by: Guenter Roeck <groeck@chromium.org>
Commit-Queue: Łukasz Bartosik <ukaszb@google.com>