fromupstream: Make marc.info work properly
I recently found a patch (message ID
20230303014930.21794-2-jing.cai@mediatek.com) that was on
marc.info. It failed to apply. It turns out that we need to append a
bit to get the raw patch out just like we do with
lore.kernel.org. This was likely always the case, but previously we
didn't support applying a patch without a "patchwork" server. Now that
we can direclty apply a patch it's nice to support marc.info.

NOTE: in the meantime, lore now has the patch, so for final testing I
ended up having to hack out the search of lore.kernel.org.

BUG=None
TEST=fromupstream.py msgid://20230303014930.21794-2-jing.cai@mediatek.com

Change-Id: If9d1b83aa0d8be0c8f5c183f5ebfb5e1c0701135
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/4308036
Tested-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>
Commit-Queue: Douglas Anderson <dianders@chromium.org>