fromupstream: Handle case when marc.info has more than one hit
Apparently there are cases where a message can show up on multiple
mailing lists that marc.info is tracking. That happened recently where
lore (for some reason) didn't have a message and then we ended up at
marc.info and saw two hits.

I couldn't find any sort of API to handle this, so we'll fall back to
just scraping the web page manually. There's no great way to do this
that's built-in to python and anything we do will be pretty brittle
anyway, so we'll just use regular expression parsing and add a special
case just for marc.info.

NOTE: in the meantime, lore now has the patch I was testing with and
so does patchwork, so for final testing I ended up having to hack out
the search of lore.kernel.org and the search of patchwork.

BUG=None
TEST=fromupstream.py msgid://8a27fb93a9d4ffdbbd86a6cdcfe57ca48041ecd9.1677752239.git.deren.wu@mediatek.com

Change-Id: I816e1e9dbb96e3c4d7bd52cd06b75e291a6145b2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/4308037
Tested-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Brian Norris <briannorris@chromium.org>
Commit-Queue: Douglas Anderson <dianders@chromium.org>