Add presubmit running CTP go unit tests
Only adding for CTP client library as that is the only code we own/not
trying to add unsolicited checks to other folks code.

The presubmit is ~10s during the first run due to downloading go CIPD
package, and ~1s afterwards. Regardless, added rudimentary logic to run
only when changes are in the CTP library and not elsewhere in the repo.

gofmt is a built in check for the presubmit already, so not adding any
linting

BUG=b:272285152
TEST=presubmit succeeds on this commit, also tried with failing tests,
and additionally changes outside the CTP library

Change-Id: Ib9778003b127856fd373c562de989868f79f30c0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/4329731
Commit-Queue: Eli Trexler <elijahtrexler@google.com>
Reviewed-by: Jack Neus <jackneus@google.com>
Tested-by: Eli Trexler <elijahtrexler@google.com>
Reviewed-by: Jaques Clapauch <jaquesc@google.com>