nebraska: Always send fw and ker versions
To better fake real Omaha, always send kernel and firmware rollback
versions, not just for ent. rollbacks. Real Omaha does so too.

Allow the kernel and firmware versions to be configured.

Also, always send updates as rollback if configured to be one, not just
if the client requested a rollback. This way, we can test that no
rollback is applied on the client side unless the rollback policy set.

BUG=b:272277145
TEST=emerge-brya update-utils
TEST=./run_unittests
TEST=Run nebraska, configure values and check for updates:
     nebraska.py --update-metadata=/var --log-file=stdout \
       --update-payloads-address file:///var
     curl -X POST -d '{"is_rollback": true,
       "rollback_prevention_image":["1.2", "1.1"],
       "rollback_prevention_stable": [["1.2", "2.3"],
         ["3.4", "4.5"], ["1.2", "2.3"],
	 ["1.2", "2.3"], ["1.2", "2.3"]]}'
       http://localhost:$(cat /run/nebraska/port)/update_config
     update_engine_client \
       --omaha_url= \
         "http://localhost:$(cat /run/nebraska/port)/update" \
       --update --skip_applying

Change-Id: I1e8a080f7bfc3b4fb2f7978f4d44352cbce3b217
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/4334574
Commit-Queue: Miriam Polzer <mpolzer@google.com>
Reviewed-by: Kyle Shimabukuro <kyleshima@chromium.org>
Reviewed-by: Jae Hoon Kim <kimjae@chromium.org>
Tested-by: Miriam Polzer <mpolzer@google.com>