cros_update_firmware: Remove deletion of cache files
Modified code to generate SRC_URI lines into ebuild to match srcuris.
Remove srcuris file once the data is transferred.

Maintain functionality of --[no]append_srcuris flags.

The flag --[no]project_srcuri does not appear to be in use currently.
This flag is a candidate for removal.

This CL should do away with the need to remove the cache directory and
the confusion it makes with all the deleted files.

BUG=b:187699294
TEST=ran tool on several boards  herobrine, trogdor, strongbad, brya

Change-Id: I5da157932a97f0851201ea5e43c0d8cfa6595d2b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/4404931
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Aamir Bohra <aamirbohra@google.com>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Tested-by: Bob Moragues <moragues@chromium.org>
Commit-Queue: Bob Moragues <moragues@chromium.org>