openwrt: Fix boot ID parsing
The boot ID used in the openwrt boot script was failing to go past
"07" since it was parsing it as an octal, and "08" is an invalid
octal resulting in a blank id from which it would then start over
at "01". This change refactors the boot script to only include the
leading zero in the boot log dirname rather than in the id var
itself so bash does not treat it as on octal.

Note: The openwrt bash shell didn't support the "<base>#" syntax for
variables so our only option was to not use the leading zeros in
integral calculations in the first place. This way, it is always
treated as a decimal.

BUG=b:271344551
TEST=successfully ran on test ap; confirmed reboot recovers and logs
as expected past boot ID "07"; validated log rotation by rebooting
device repeatedly

Cq-Depend: 4416404
Change-Id: I1fea4703a340cc16f34acb1a12e4c8070665a18b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/4416403
Commit-Queue: Jared Bennett <jaredbennett@google.com>
Tested-by: Jared Bennett <jaredbennett@google.com>
Reviewed-by: Billy Zhao <billyzhao@chromium.org>