openwrt: Fix last boot failure check
The LAST_CHECK_WAS_FAILURE value was meant to be the exit code
of the grep call, but was actually set to its output (which is
always empty). This resulted in the comparison before the reboot
to always throw an error since it would be trying to compare an
empty string to a number, thus never reaching the reboot step.

This change correctly sets LAST_CHECK_WAS_FAILURE to the exit
code of the grep call. The image version was also updated.

BUG=b:271344551
TEST=manually confirmed updated bash commands and if statements work
as expected on an openwrt device

Change-Id: I83add73b4de857ea4956b036e79859b64560502b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/dev-util/+/4435991
Auto-Submit: Jared Bennett <jaredbennett@google.com>
Reviewed-by: Billy Zhao <billyzhao@chromium.org>
Tested-by: Jared Bennett <jaredbennett@google.com>
Commit-Queue: Billy Zhao <billyzhao@chromium.org>