Setup git-ps to do stacked reviews on github.
This will allow us to have stacked reviews like in gerrit more easily
and send a chain.

We can use front ends for github like https://reviewstack.dev/ (or many
others, but this is a nice one maintained by the folks at facebook) to
see these in a nice way like gerrit provides.

<!-- ps-id: 9450555b-6b25-4286-9e57-e2e1cbbe568e -->