remove webUSB (#137)
This PR removes "Start (webUSB)" button to make the codes simpler.
It was implemented to support MacOS, which doesn't support communication via WebSerial. But for now, we are focusing on ChromeOS as a workstation and DUT, so it can be removed.
If MacOS is needed for the future, you can reproduce the codes from git log.