Add annotation (#140)
Add annotation in the power graph when the "start" or "end" signal comes in DUT console.
X-axis value is expressed as a relative time from the start of measurement
to avoid adding annotation when copying shell script to DUT. Also the script is encoded in base64 to include annotation information to export the JSON file. The data format is defined in the design doc.