Force deploy to avoid failure caused by USE mismatch warnings
Force deploy when running `cros deploy` via `lium deploy`.
`cros deploy` started checking if USE flags are matched when deploying
and building, and when it is not matched, it asks users to proceed or
not (and... the command itself will succeed even if the stdin sends just
an EOF, which is not good...), which causes a false success on the lium
side. To avoid this, for now, calling `cros deploy` with `--force` flag
to continue without that prompt.