219 cro3 flash recovery should use signed image on usb (#239)
* flash: Make --recovery requires --usb

Since --recovery flashes a USB recovery image, it should be written
in USB flash. If we directly write it to DUT, the DUT will not boot.

* flash: Use signed reocvery image instead of normal recovery image

When we recover DUT from bricked state, we need to use signed image
instead of unsigned one. Use "signed" recovery image for making a
recovery USB flash.